backend/requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
python-decouple==3.8
pandas==2.1.4
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
django-extensions==3.2.3
